Removed functors and functions from header as detail implementation
